feat: make it so hidden roles and assignments are visible to everyone
I am changing the `is_hidden` property to make it so certain roles are hidden manually by the frontend, in case, for example, a user has 3 important roles to focus on and 10 others which aren't really important. I don't think there is really a good security reason to make roles hidden.