fix: fix permissions for a user updating a profile
Introduces a view called `profiles_for_utag` which lists all the profiles that have a particular `utag`.

Introduces a view called `profiles_for_urole` which lists all the profiles that have a particular 'urole'.

Refactors permission logic for profile updates to use row-level security based on actions defined in `permission_action`.

This change replaces the previous policy, which allowed users to update their own profile, with a new policy that checks if the user has the `edit_profile` permission. It also introduces new views to determine which profiles a user can modify based on their roles and tags.

The previous update permissions have been revoked to enforce the new policy.